implemented details sorting
